Revert "feat: 마이페이지 관련 조회 기능들 구현"
