+++ baemin/src/main/java/com/sist/baemin/store/domain/ReviewEntity.java
@@ -1,13 +1,20 @@
 package com.sist.baemin.store.domain;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 import com.sist.baemin.order.domain.OrderEntity;
 import com.sist.baemin.user.domain.UserEntity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -46,6 +53,4 @@ public class ReviewEntity {
     
     @Column(name = "createdAt")
     private LocalDateTime createdAt;
-
-
 }
