+++ baemin/src/main/java/com/sist/baemin/user/domain/CustomUserDetails.java
@@ -1,6 +1,5 @@
 package com.sist.baemin.user.domain;
 
-import lombok.ToString;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -8,7 +7,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
 import java.util.List;
 
-@ToString
 public class CustomUserDetails implements UserDetails {
     private final UserEntity user;
 
@@ -16,9 +14,6 @@ public class CustomUserDetails implements UserDetails {
         this.user = user;
     }
 
-    public Long getUserId() {
-        return user.getUserId();
-    }
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return List.of(new SimpleGrantedAuthority("ROLE_USER"));
