+++ baemin/src/main/java/com/sist/baemin/user/dto/UserAddressDto.java
@@ -1,12 +1,14 @@
 package com.sist.baemin.user.dto;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@ToString
 public class UserAddressDto {
     private Long id;
     private String alias;
