+++ baemin/src/main/java/com/sist/baemin/user/dto/UserProfileDto.java
@@ -1,19 +1,18 @@
 package com.sist.baemin.user.dto;
 
-import lombok.*;
-
-import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@ToString
 public class UserProfileDto {
     private String nickname;
     private String realName;
     private String email;
     private String profileImageUrl;
-    private Long couponCount;
-    private Long point;
-}
+    private String defaultAddress;
+}
