--- baemin/src/main/java/com/sist/baemin/user/repository/UserAddressRepository.java
@@ -1,12 +1,0 @@
-package com.sist.baemin.user.repository;
-
-import com.sist.baemin.user.domain.UserAddressEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface UserAddressRepository extends JpaRepository<UserAddressEntity, Long> {
-    List<UserAddressEntity> findByUser_Email(String email);
-}
