+++ baemin/src/main/java/com/sist/baemin/user/repository/UserCouponRepository.java
@@ -11,7 +11,6 @@ import com.sist.baemin.user.domain.UserCouponEntity;
 @Repository
 public interface UserCouponRepository extends JpaRepository<UserCouponEntity, Long> {
     List<UserCouponEntity> findByExpiresAtBefore(LocalDateTime cutoff);
-    long countByUser_EmailAndIsUsed(String email, boolean isUsed);
 }
