--- baemin/src/main/resources/templates/html/mypage-profile.html
@@ -1,171 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>마이배민 - 내 정보 수정</title>
-    <link rel="stylesheet" href="/css/mypage.css">
-</head>
-<body>
-<div class="baemin-page">
-    <header class="appbar">
-        <a class="icon-btn back" href="/api/mypage" aria-label="뒤로">←</a>
-        <h1 class="appbar-title">마이배민</h1>
-    </header>
-
-    <main class="content">
-        <section class="view active">
-            <div class="page-title">내 정보 수정</div>
-            <div class="profile-edit">
-                <div class="avatar large">
-                    <img id="profileImg"
-                         th:src="${(user?.profileImageUrl) ?: 'https://via.placeholder.com/96x96/EEE/888?text=👤'}"
-                         alt="프로필 이미지">
-                    <button class="camera-btn" onclick="openImageUpload()">📷</button>
-                </div>
-
-                <ul class="list">
-                    <li class="list-item" id="nicknameItem" role="button">
-                        <span class="label">닉네임</span>
-                        <span class="value"
-                              th:text="${#strings.isEmpty(user?.nickname) ? '닉네임' : user.nickname}">닉네임
-                        </span>
-                        <span class="chevron">›</span>
-                    </li>
-                    <li class="list-item">
-                        <span class="label">이름</span>
-                        <span class="value"
-                              th:text="${#strings.isEmpty(user?.realName) ? '이름' : user.realName}">이름</span>
-                        <span class="chevron">›</span>
-                    </li>
-                    <li class="list-item">
-                        <span class="label">대표 이메일</span>
-                        <span class="value"
-                              th:text="${#strings.isEmpty(user?.email) ? '이메일' : user.email}">이메일</span>
-                        <span class="chevron">›</span>
-                    </li>
-                    <li class="list-item muted" role="button">
-                        <span class="label">휴대폰 번호 변경</span>
-                        <span class="chevron">›</span>
-                    </li>
-                </ul>
-
-                <ul class="list">
-                    <li class="list-item">
-                        <span class="label">로그인 기기 관리</span>
-                        <span class="helper">내 아이디로 로그인 된 기기를 관리할 수 있어요</span>
-                        <span class="chevron">›</span>
-                    </li>
-                </ul>
-
-                <ul class="list">
-                    <li class="list-item">
-                        <span class="label">연동된 소셜 계정</span>
-                        <span class="value">💬</span>
-                        <span class="chevron">›</span>
-                    </li>
-                </ul>
-
-                <div class="bottom-links">
-                    <button class="text-btn" onclick="logoutWithService()">로그아웃</button>
-                    <span class="divider">|</span>
-                    <button class="text-btn danger" onclick="deleteAccount()">회원탈퇴</button>
-                </div>
-            </div>
-        </section>
-    </main>
-
-    <nav class="bottombar">
-        <a class="bottom-item" href="/api/main">
-            <span>🏠</span>
-            <span>홈</span>
-        </a>
-        <a class="bottom-item" href="#">
-            <span>🛍️</span>
-            <span>장보기·쇼핑</span>
-        </a>
-        <a class="bottom-item" href="#">
-            <span>❤</span>
-            <span>찜</span>
-        </a>
-        <a class="bottom-item" href="#">
-            <span>🧾</span>
-            <span>주문내역</span>
-        </a>
-        <a class="bottom-item active" href="/api/mypage">
-            <span>🙂</span>
-            <span>마이배민</span>
-        </a>
-    </nav>
-</div>
-
-<input type="file" id="profileImageInput" accept="image/*" style="display:none" onchange="uploadProfileImage()">
-
-<script>
-    function openImageUpload() { document.getElementById('profileImageInput').click(); }
-
-    async function uploadProfileImage() {
-        const fileInput = document.getElementById('profileImageInput');
-        const file = fileInput.files[0];
-        if (!file) return;
-        if (file.size > 5 * 1024 * 1024) { alert('파일 크기는 5MB 이하만 가능합니다.'); return; }
-        const allowed = ['image/jpeg','image/jpg','image/png'];
-        if (!allowed.includes(file.type)) { alert('JPG/PNG 형식만 가능합니다.'); return; }
-
-        const profileImg = document.getElementById('profileImg');
-        const originalSrc = profileImg.src;
-        profileImg.style.opacity = '0.5';
-
-        const formData = new FormData();
-        formData.append('profileImage', file);
-        try {
-            const res = await fetch('/api/user/profile-image', { method: 'POST', credentials: 'include', body: formData });
-            const result = await res.json();
-            if (result.resultCode === 200) {
-                const newUrl = result.data.imageUrl + '?t=' + Date.now();
-                profileImg.src = newUrl;
-                profileImg.style.opacity = '1';
-            } else {
-                alert('업로드 실패: ' + result.message);
-                profileImg.src = originalSrc; profileImg.style.opacity = '1';
-            }
-        } catch (e) {
-            console.error(e); alert('네트워크 오류가 발생했습니다.');
-            profileImg.src = originalSrc; profileImg.style.opacity = '1';
-        } finally {
-            fileInput.value = '';
-        }
-    }
-
-    // 닉네임 편집 페이지로 이동
-    (function(){
-        const el = document.getElementById('nicknameItem');
-        if (el) {
-            el.addEventListener('click', () => {
-                window.location.href = '/api/mypage/profile/nickname';
-            });
-        }
-    })();
-
-    function logoutWithService() {
-        if (confirm('카카오계정과 함께 로그아웃 하시겠습니까?')) {
-            const kakaoLogoutUrl = "https://kauth.kakao.com/oauth/logout" +
-                "?client_id=9332367d804b05aa4921d0ddd1c788cb" +
-                "&logout_redirect_uri=http://localhost:8080/api/logout";
-            window.location.href = kakaoLogoutUrl;
-        }
-    }
-
-    async function deleteAccount() {
-        if (!confirm('정말로 회원 탈퇴하시겠습니까?')) return;
-        if (!confirm('마지막 확인입니다. 정말로 탈퇴하시겠습니까?')) return;
-        try {
-            const res = await fetch('/api/user/account', { method: 'DELETE', credentials: 'include' });
-            const result = await res.json();
-            if (result.resultCode === 200) window.location.href = '/';
-            else alert('회원 탈퇴 실패: ' + result.message);
-        } catch (e) { alert('회원 탈퇴 중 오류'); }
-    }
-</script>
-</body>
-</html>
