--- baemin/src/main/resources/templates/html/mypage-reviews.html
@@ -1,126 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>마이배민 - 리뷰관리</title>
-    <link rel="stylesheet" href="/css/mypage.css">
-</head>
-<body>
-<div class="baemin-page">
-    <header class="appbar">
-        <a class="icon-btn back" href="/api/mypage" aria-label="뒤로">←</a>
-        <h1 class="appbar-title">마이배민</h1>
-    </header>
-
-    <main class="content">
-        <section class="view active">
-            <div class="page-title">리뷰관리</div>
-            <div class="tabbar">
-                <button class="tab active">배달·픽업</button>
-                <button class="tab" disabled>장보기·쇼핑</button>
-            </div>
-            <div class="review-count">내가 쓴 총 리뷰 <b id="reviewCount" th:text="${(reviews != null) ? #lists.size(reviews) + '개' : '0개'}">0개</b></div>
-
-            <div id="reviewList" class="review-list">
-                <article class="review-card" th:each="r : ${reviews}">
-                    <header class="review-header">
-                        <div>
-                            <div class="store-name" th:text="${r.storeName}">가게명</div>
-                            <div class="meta" th:text="${#strings.repeat('★', r.rating)}">★★★★★</div>
-                        </div>
-                        <button class="small-btn" th:attr="data-id=${r.reviewId}">삭제</button>
-                    </header>
-                    <p class="review-text" th:text="${r.content}">리뷰 내용</p>
-                    <div class="review-image" th:if="${r.reviewImage} != null"
-                         th:style="${'background-image:url(' + r.reviewImage + ')'}"></div>
-                </article>
-
-                <p class="empty" th:if="${reviews == null or #lists.isEmpty(reviews)}">등록된 리뷰가 없습니다.</p>
-            </div>
-        </section>
-    </main>
-
-    <nav class="bottombar">
-        <a class="bottom-item" href="/api/main">
-            <span>🏠</span>
-            <span>홈</span>
-        </a>
-        <a class="bottom-item" href="#">
-            <span>🛍️</span>
-            <span>장보기·쇼핑</span>
-        </a>
-        <a class="bottom-item" href="#">
-            <span>❤</span>
-            <span>찜</span>
-        </a>
-        <a class="bottom-item" href="#">
-            <span>🧾</span>
-            <span>주문내역</span>
-        </a>
-        <a class="bottom-item active" href="/api/mypage">
-            <span>🙂</span>
-            <span>마이배민</span>
-        </a>
-    </nav>
-</div>
-
-<script>
-    document.addEventListener('DOMContentLoaded', () => {
-        const updateCount = () => {
-            const el = document.getElementById('reviewCount');
-            if (!el) return;
-            const count = document.querySelectorAll('#reviewList .review-card').length;
-            el.textContent = count + '개';
-        };
-
-        updateCount();
-
-        const listEl = document.getElementById('reviewList');
-        if (!listEl) return;
-
-        listEl.addEventListener('click', async (e) => {
-            const btn = e.target.closest('.small-btn');
-            if (!btn || !listEl.contains(btn)) return;
-
-            const reviewId = btn.getAttribute('data-id');
-            if (!reviewId) return;
-
-            if (!confirm('이 리뷰를 삭제하시겠어요?')) return;
-
-            try {
-                const res = await fetch(`/api/user/reviews/${reviewId}`, {
-                    method: 'DELETE',
-                    credentials: 'include',
-                    headers: { 'Accept': 'application/json' }
-                });
-
-                let result = null;
-                try { result = await res.json(); } catch (_) {}
-
-                if (res.ok) {
-                    const card = btn.closest('.review-card');
-                    if (card) card.remove();
-                    updateCount();
-
-                    // 비어있을 때 안내 추가
-                    const remaining = document.querySelectorAll('#reviewList .review-card').length;
-                    if (remaining === 0) {
-                        const empty = document.createElement('p');
-                        empty.className = 'empty';
-                        empty.textContent = '등록된 리뷰가 없습니다.';
-                        listEl.appendChild(empty);
-                    }
-
-                    alert('삭제되었습니다.');
-                } else {
-                    alert((result && result.message) ? result.message : '삭제에 실패했습니다.');
-                }
-            } catch (err) {
-                alert('네트워크 오류로 삭제에 실패했습니다.');
-            }
-        });
-    });
-</script>
-</body>
-</html>
